Move the `test` module out into its own file
